+++ src/vcastTestInterface.ts
@@ -16,7 +16,6 @@ import {
 } from "./editorDecorator";
 
 import { 
-  buildEnvironmentCallback, 
   showSettings, 
   updateDataForEnvironment
  } from "./helper";
@@ -31,7 +30,7 @@ import {
   getEnviroPathFromID, 
   testNodeType } from "./testData";
 import {
-  executeCommand,
+  executeCommandSync,
   executeVPythonScript,
   forceLowerCaseDriveLetter,
   getChecksumCommand,
@@ -77,9 +76,9 @@ function getChecksum(filePath: string) {
         path.dirname(filePath)
       );
     else
-      commandOutputString = executeCommand(
-        `${checksumCommand} ${filePath}`
-      ).stdout;
+      commandOutputString = executeCommandSync(
+        `${checksumCommand} ${filePath}`,
+        process.cwd()).stdout;
 
     // convert the to a number and return
     // this will crash if something is wrong with the result
@@ -825,21 +824,50 @@ export async function newTestScript(testNode: testNodeType) {
   );
 }
 
+
+function quote (name:string){
+  // if name contains <<COMPOUND>>, <<INIT>> or parenthesis
+  // we need to quote the name so that the shell does not interpret it.
+
+  if (
+    name.includes("<") ||
+    name.includes(">") ||
+    name.includes("(") ||
+    name.includes(")")
+  ) {
+    return '"' + name + '"';
+  }
+  else 
+    return name;
+}
+
+export function getClicastArgsFromTestNodeAsList(testNode: testNodeType):string[] {
+  // this function will create the enviro, unit, subprogram, and test
+  // arguments as a list, since spawn for example requires an arg list.
+
+  let returnList = [];
+  returnList.push(`-e${testNode.enviroName}`);
+  if (testNode.unitName.length > 0 && testNode.unitName != "not-used")
+    returnList.push (`-u${testNode.unitName}`);
+
+  // we need the quotes on the names to handle <<COMPOUND>>/<<INIT>>/parenthesis
+  if (testNode.functionName.length > 0)
+    returnList.push (`-s${quote (testNode.functionName)}`);
+  if (testNode.testName.length > 0)  {
+    const nameToUse = testNode.testName.replace (compoundOnlyString,"");
+    returnList.push (`-t${quote(nameToUse)}`);
+  }
+
+  return returnList;
+}
+
+
 export function getClicastArgsFromTestNode(testNode: testNodeType) {
   // this function will create the enviro, unit, subprogram,
-  // and test arg string for clicast
+  // and test arg string for clicast calls that need a arg string
 
-  let returnString = "";
-  returnString = `-e${testNode.enviroName}`;
-  if (testNode.unitName.length > 0 && testNode.unitName != "not-used")
-    returnString += ` -u${testNode.unitName}`;
-
-  // we need the quotes on the names to handle <<COMPOUND>> and <<INIT>>
-  if (testNode.functionName.length > 0)
-    returnString += ` -s"${testNode.functionName}"`;
-  if (testNode.testName.length > 0) returnString += ` -t"${testNode.testName.replace (compoundOnlyString,"")}"`;
-
-  return returnString;
+  const argList = getClicastArgsFromTestNodeAsList (testNode);
+  return argList.join (" ");
 }